Refactor navigation components into navBar directory
Moved and split Navigation, AccountDetails, ColorfulRequestPrintButton, Footer, NavbarDesktop, NavbarLink, NavbarMobile, and ScreenDimmer components into a dedicated navBar directory for improved modularity and maintainability. Updated layout.tsx imports to use new component locations.
